+++ lib/warpath/expression/parser.ex
@@ -1,11 +1,11 @@
-defmodule Warpath.Parser do
+defmodule Warpath.Expression.Parser do
   @moduledoc false
 
-  alias Warpath.ParserError
+  alias Warpath.Expression.ParserError
 
   def parse(tokens) when is_list(tokens) do
     tokens
-    |> :parser.parse()
+    |> :warpath_parser.parse()
     |> case do
       {:ok, _} = expression_tokens ->
         expression_tokens